add pundit

+++ app/controllers/posts_controller.rb
@@ -1,7 +1,7 @@
 class PostsController < ApplicationController
   before_action :set_post, only: [:show, :edit, :update, :destroy]
   before_action :authenticate_user!, except: [:index, :show]
-  before_action :validate_post_owner, only: [:edit, :update, :destroy]
+  
 
   def index
     @posts = Post.includes(:categories, :user, :region, :province).page(params[:page]).per(5)
@@ -26,10 +26,12 @@ class PostsController < ApplicationController
   def show; end
 
   def edit
+    authorize @post, :edit?, policy_class: PostPolicy
     session[:return_to] = request.referer
   end
 
   def update
+    authorize @post, :update?, policy_class: PostPolicy
     if @post.update(post_params)
       flash[:notice] = 'Post updated successfully'
       redirect_to session.delete(:return_to)
@@ -40,19 +42,13 @@ class PostsController < ApplicationController
   end
 
   def destroy
+    authorize @post, :destroy?, policy_class: PostPolicy
     @post.destroy
     flash[:notice] = 'Post destroyed successfully'
     redirect_to posts_path
   end
 
   private
-
-  def validate_post_owner
-    unless @post.user == current_user
-      flash[:notice] = 'This post does not belong to you'
-      redirect_to posts_path
-    end
-  end
 
   def set_post
     @post = Post.find(params[:id])
